refactor: only use msalInstance within MsalProvider
