+++ src/auth/apiClient.ts
@@ -4,6 +4,7 @@ import { apiConfig } from './authConfig'
 
 export const apiClient = createClient<paths>({
   baseUrl: apiConfig.baseUrl,
+  // headers: new Headers({ Authorization: `Bearer ${token}` }),
 })
 
 export type ApiClient = typeof apiClient
