+++ src/hooks/useGetAnalogueModels.tsx
@@ -1,14 +1,6 @@
 import { useQuery } from '@tanstack/react-query'
 import { useApiClient } from '../context/ApiClientProvider'
-import { useAccessToken } from './useToken'
-
-// type UseQueryOptions<T> = Params<T> &
-//   RequestBody<T> & {
-//     options?: T
-//     reactQuery?: {
-//       enabled: boolean
-//     }
-//   }
+import { useAccessToken } from './useAccessToken'
 
 const ANALOGUEMODELS_KEY = '/api/analoguemodels'
 
@@ -25,7 +17,5 @@ export function useGetAnalogueModels() {
           headers: new Headers({ Authorization: `Bearer ${token}` }),
         })
         .then((response) => response.data?.data)
-    // React Query expects errors to be thrown to show a message
-    // throw new Error(error.message)
   )
 }
