--- src/hooks/useToken.tsx
@@ -1,14 +1,0 @@
-import { useEffect, useState } from 'react'
-import { fetchAccessToken } from '../auth/msalClient'
-
-export const useAccessToken = () => {
-  const [accessToken, setAccessToken] = useState<string | undefined>()
-
-  useEffect(() => {
-    if (!accessToken) {
-      fetchAccessToken().then((token) => setAccessToken(token))
-    }
-  }, [accessToken])
-
-  return accessToken
-}
